api/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
redis